Move home page to app router
