+++ frontend/src/app/[locale]/page.tsx
@@ -1,0 +1,31 @@
+import BetaAlert from "src/components/BetaAlert";
+import PageSEO from "src/components/PageSEO";
+import Hero from "src/components/Hero";
+import IndexGoalContent from "src/components/content/IndexGoalContent";
+import ProcessAndResearchContent from "src/components/content/ProcessAndResearchContent";
+import { Metadata } from "next";
+import { useTranslations } from "next-intl";
+import { getTranslations } from "next-intl/server";
+
+export async function generateMetadata() {
+  const t = await getTranslations({ locale: "en" });
+  const meta: Metadata = {
+    title: t("Index.page_title"),
+    description: t("Index.meta_description"),
+  };
+  return meta;
+}
+
+export default function Home() {
+  const t = useTranslations("Index");
+
+  return (
+    <>
+      <PageSEO title={t("page_title")} description={t("meta_description")} />
+      <Hero />
+      <BetaAlert />
+      <IndexGoalContent />
+      <ProcessAndResearchContent />
+    </>
+  );
+}
