+++ frontend/src/components/content/IndexGoalContent.tsx
@@ -1,11 +1,12 @@
-import { useTranslation } from "next-i18next";
-import Link from "next/link";
-import { Button, Grid, Icon } from "@trussworks/react-uswds";
+import { useTranslations } from "next-intl";
 
+import Link from "next/link";
+import { Button, Grid } from "@trussworks/react-uswds";
+import { USWDSIcon } from "../USWDSIcon";
 import ContentLayout from "src/components/ContentLayout";
 
 const IndexGoalContent = () => {
-  const { t } = useTranslation("common", { keyPrefix: "Index" });
+  const t = useTranslations("Index");
 
   return (
     <ContentLayout
@@ -18,9 +19,9 @@ const IndexGoalContent = () => {
         <Link href="/newsletter" passHref>
           <Button className="margin-bottom-4" type="button" size="big">
             <span className="margin-right-5">{t("goal.cta")}</span>
-            <Icon.ArrowForward
-              className="text-middle margin-left-neg-4"
-              size={4}
+            <USWDSIcon
+              name="arrow_forward"
+              className="usa-icon usa-icon--size-4 text-middle margin-left-neg-4"
               aria-label="arrow-forward"
             />
           </Button>
