+++ frontend/stories/pages/Index.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta } from "@storybook/react";
-import Index from "src/pages/index";
+import Index from "src/app/[locale]/page";
 
 const meta: Meta<typeof Index> = {
   title: "Pages/Home",
